Update neomutt.rb
With a stock install of linuxbrew, the configure step fails since it can't find xlstproc and krb5-config. Adding these as build dependencies allows the configure step to complete.

The sasl flag fails to find sasl/sasl.h, which is currently not provided by any linuxbrew packages to my knowledge. I had to remove the --sasl flag for the configure step to complete.